Rename the storybook entry file
