+++ config.js
@@ -1,7 +1,7 @@
 import { configure } from '@storybook/react';
 
 function loadStories() {
-  require('./stories/test.js');
+  require('./stories/index.js');
 }
 
 configure(loadStories, module);
